feat(concerto): Add regex validation and explicit subclass selection
